Split tests into two files, extract commons
